+++ src/video.h
@@ -43,7 +43,7 @@ class Video
 		virtual string toString() = 0;
 		virtual string toCsvString() = 0;
 		Video* operator + (Video& otherVideo);
-		void addReview(int review);
+		void addReview(float review);
 };
 
 #endif
